cs 6680 - #2 - improve problem 3 of assignment 1
I timed various ways of calculating the minimum, maximum, and mean. I selected
the fastest method for the minimum and maximum. The mean methods were all
comparable in speed, so selected the easiest to read. There is only one way to
correctly calculate the median, so efficiency and readability are out the door.