cs 6680 - #10 - change uav_trainer to a function
uav_trainer.m is now a function, which trains the keypoint classification SVM,
and saves it to disk. a new file, uav_car_counter.m, is the entry point for the
program. it checks for the keypoint SVM file. if it exists, it loads the SVM.
otherwise, it trains a new SVM.